Removed doNotTrack. Added initial support for lastGhost to session.
Removed doNotTrack functionality since users are accessing public rooms anyway. Will add requirement for providing list of roomIds to the server when calling roomlist to compensate for this. That way you must specifically know the roomId in order to get data for it which will allow private rooms to be protected. Added initial support for storing user's last ghost model data in the session so it can be referenced later and sent when calling roomlist